% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gclassificatin_func.R
\name{extract_node_v_helper}
\alias{extract_node_v_helper}
\title{A function that extracts the nodes of a data.tree object where the given
attribute is not NA.
A helper function that is not called by the user}
\usage{
extract_node_v_helper(data_tree, attribute)
}
\arguments{
\item{data_tree}{A data.tree object}

\item{attribute}{Character string that specifies the attribute of interest}
}
\description{
A function that extracts the nodes of a data.tree object where the given
attribute is not NA.
A helper function that is not called by the user
}
\examples{
test = extract_node_v(base_tree_a, attribute = "station_id")
}
\author{
Stefan Lüdtke
}
