% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gclassificatin_func.R
\name{link_root_node}
\alias{link_root_node}
\title{A function that takes two data.tree object, where reduced_tree is a subtree
of full_tree.
A helper function that is not called by the user}
\usage{
link_root_node(full_tree, reduced_tree)
}
\arguments{
\item{full_tree}{data.tree object that provides the source}

\item{reduced_tree}{data.tree object as a subtree of full_tree}
}
\value{
data.tree object
}
\description{
A function that takes two data.tree object, where reduced_tree is a subtree
of full_tree.
A helper function that is not called by the user
}
\examples{
test = link_root_node(base_tree_b, reduced_tree_b)
}
\author{
Stefan Lüdtke
}
