% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gclassificatin_func.R
\name{parse_outlet}
\alias{parse_outlet}
\title{How to document functions.}
\usage{
parse_outlet(node, func_station_id = NA)
}
\arguments{
\item{text}{Message to print.}
}
\description{
This is an example showing how to document functions to be included in the package.
}
\examples{
parse_outlet('Hello world!')
}
\seealso{
\url{http://r-pkgs.had.co.nz/man.html}
}
\author{
Stefan Lüdtke
}
