% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gclassificatin_func.R
\name{reduce_tree_helper}
\alias{reduce_tree_helper}
\title{A helper function that is not called by the user}
\usage{
reduce_tree_helper(x, data_tree, attribute)
}
\arguments{
\item{x}{A node name to start with}

\item{data_tree}{A tree to operate on}
}
\value{
a data frame with 1 row and the node name from the next ancestor
}
\description{
A helper function that is not called by the user
}
\author{
Stefan Lüdtke
}
